Enum crndm::stm::LogEnum

pub enum LogEnum {
    DataLog(u64, u64, usize),
    DropOnAbort(u64, usize),
    DropOnCommit(u64, usize),
    DropOnFailure(u64, usize),
    UnlockOnCommit(u64),
    None,
}

Log Types

Variants

DataLog(u64, u64, usize)

(src, log, len): An undo log of slice src..src+len kept in log..log+len.

DropOnAbort(u64, usize)

(u64, usize): Similar to DropOnFailure except that it drops the allocation when the high-level transaction is aborted. This is useful for temporarily unowned allocations, such as slices, because they are not deallocated via RAII.

DropOnCommit(u64, usize)

(src, len): A drop log indicating that slice src..src+len should drop on commit, useful for in-transactional drop functions.

DropOnFailure(u64, usize)

(src, len): A drop log indicating that slice src..src+len should drop on failure, useful for high-level allocation.

UnlockOnCommit(u64)

Unlocks a Mutex on transaction commit.

None